Implemented OntologyTermService read service methods
